giroffsets: Use size types for size and alignments
Note that for alignments we should actually use size_t, but this would
imply some refactors since this value can be acually set to -1 in our
implementation.

So we use gssize for now, that at least on gcc is defined.

+++ girepository/girnode-private.h
@@ -261,8 +261,8 @@ struct _GIIrNodeInterface
   GList *interfaces;
   GList *prerequisites;
 
-  int alignment;
-  int size;
+  gssize alignment;
+  size_t size;
 
   GList *members;
 };
@@ -311,8 +311,8 @@ struct _GIIrNodeBoxed
   char *gtype_name;
   char *gtype_init;
 
-  int alignment;
-  int size;
+  gssize alignment;
+  size_t size;
 
   GList *members;
 };
@@ -334,8 +334,8 @@ struct _GIIrNodeStruct
   char *copy_func;
   char *free_func;
 
-  int alignment;
-  int size;
+  gssize alignment;
+  size_t size;
 
   GList *members;
 };
@@ -355,8 +355,8 @@ struct _GIIrNodeUnion
   char *copy_func;
   char *free_func;
 
-  int alignment;
-  int size;
+  gssize alignment;
+  size_t size;
 
   int discriminator_offset;
   GIIrNodeType *discriminator_type;
